Multi-echo tried but no idea of optimal reward
So avg-reward calculation is not proper as the rewards here not weighted according to target size as discussed. Target size can vary from 1-4. Need to find a way to normalize each reward for each target size, to get proper avg-reward and better idea of what the optimal idea should be.