Multi-echo upto 3 working
I think in this case, even though the target size is variable, avg optimal reward can be determined by (1+2+3)/3..and can be scaled up accordingly. Reason being if considering the best case scenario of choosing 1 or 2 or 3 equally likely, the prob for each is 1/3. So according the reward can become 1/3,2/3 and 3/3 and the optimal will be 2 all epochs, considering episodes. Best case