Modified reward, huge improvement
For reward - all zeros get -( len(target)+1 ) to make it the worst situation to be. Max reward can be 1.8 (+2 for correct elements - 0.1*2 for 2 zeros). Range is (-5,1.8) for rewards, and rewards aren't really proportional. So 1.6+/1.8 avg rewards should be good performance as even one -5 reward in 300 episodes will reduce more from the max total of 1.8 as avg reward